+++ docs_src/tutorials/defining_dependencies_products_dependencies_py310.py
@@ -1,6 +1,8 @@
 from pathlib import Path
 from typing import Annotated
 
+import matplotlib.pyplot as plt
+import pandas as pd
 from my_project.config import BLD
 from pytask import Product
 
@@ -9,4 +11,10 @@ def task_plot_data(
     path_to_data: Path = BLD / "data.pkl",
     path_to_plot: Annotated[Path, Product] = BLD / "plot.png",
 ) -> None:
-    ...
+    df = pd.read_pickle(path_to_data)
+
+    _, ax = plt.subplots()
+    df.plot(x="x", y="y", ax=ax, kind="scatter")
+
+    plt.savefig(path_to_plot)
+    plt.close()